move command to common from contributions
